backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0
redis==5.2.0
celery==5.4.0
openai==1.58.1
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf==5.1.0
tiktoken==0.8.0
python-multipart==0.0.19
boto3==1.35.81
sentry-sdk==2.19.2
httpx==0.28.1
email-validator==2.2.0
pytest==8.3.4
